Add files via upload
Added ".0" to the SourceRef to indicate the start of the verse before the first subverse (which was sometimes missing)
Simplified Est. refs, eg (Hebrew=10:3 • Latin=10:3, 3a-k[=10:4-13; 11:1] • Greek=10:3, 3a-k[=10:4-13; 11:1 (F:1-11)]) to (Hebrew=10:3 • Latin=10:3; F:1-11[=10:4-13; 11:1] • Greek=10:3, 3a-k[=10:4-13; 11:1])  and (Hebrew= -- • Latin=1:1f[=11:6] • Greek=1:1f[=11:7 (A:6)]) to (Hebrew= -- • Latin=A:6[=11:6] • Greek=1:1f[=11:7])
Added Esg and filled in varieties of Est.