Add files via upload
I've had a change of heart wrt Esther
The extra material in Esther can be inserted into the Standard versification as subverses,
or it can be added as extra chapters 11-15
or it can be added as extra chapters called A-F (which unfortunately don't match the extra numbered chapters). 
Initially I had the Callouts and Notes often referring to the extra chaps A-F. 
I've now changed that to extra chaps 11-15 (except in Callouts where the original has chaps A-F) because I think this is more common and certainly more understandable. I've also upgraded the mappings so that all three options are shown, eg (Hebrew= -- • Latin=11:7/A:6 • Greek=1:1f)

I've also added new Phrases to translate: 
1) %In some Bibles only the start of this verse is present%.
  - This replaces notes such as: 
   "%Normally in this Bible this verse includes words that are at% 11:19, 21-22, 24, 26-28"
2) %In some Bibles this verse contains extra text%.
  - This is added to Bibles (eg Hebrew and Protestant Bibles) that do not have these extra subverses